+++ app/Http/Requests/PriceCategoryRequest.php
@@ -4,9 +4,10 @@ declare(strict_types=1);
 
 namespace Kami\Cocktail\Http\Requests;
 
+use Kami\Cocktail\Rules\ValidCurrency;
 use Illuminate\Foundation\Http\FormRequest;
 
-class IngredientRequest extends FormRequest
+class PriceCategoryRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -27,12 +28,7 @@ class IngredientRequest extends FormRequest
     {
         return [
             'name' => 'required',
-            'strength' => 'numeric',
-            'prices' => 'array',
-            'prices.*.price' => 'required|gte:0|decimal:0,2',
-            'prices.*.amount' => 'required|numeric|gte:0',
-            'prices.*.units' => 'required',
-            'prices.*.price_category_id' => 'int|required',
+            'currency' => ['required', 'size:3', new ValidCurrency()],
         ];
     }
 }