+++ app/Models/CocktailFavorite.php
@@ -6,12 +6,9 @@ namespace Kami\Cocktail\Models;
 
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
-use Illuminate\Database\Eloquent\Factories\HasFactory;
 
 class CocktailFavorite extends Model
 {
-    use HasFactory;
-
     /**
      * @return BelongsTo<Cocktail, CocktailFavorite>
      */