+++ app/Models/CocktailIngredientSubstitute.php
@@ -6,12 +6,9 @@ namespace Kami\Cocktail\Models;
 
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
-use Illuminate\Database\Eloquent\Factories\HasFactory;
 
 class CocktailIngredientSubstitute extends Model
 {
-    use HasFactory;
-
     /**
      * @return BelongsTo<CocktailIngredient, CocktailIngredientSubstitute>
      */