+++ app/Models/UserIngredient.php
@@ -6,12 +6,9 @@ namespace Kami\Cocktail\Models;
 
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
-use Illuminate\Database\Eloquent\Factories\HasFactory;
 
 class UserIngredient extends Model
 {
-    use HasFactory;
-
     public $timestamps = false;
 
     /**