+++ app/Providers/AppServiceProvider.php
@@ -2,6 +2,7 @@
 
 namespace Kami\Cocktail\Providers;
 
+use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\ServiceProvider;
 
 class AppServiceProvider extends ServiceProvider
@@ -23,6 +24,6 @@ class AppServiceProvider extends ServiceProvider
      */
     public function boot()
     {
-        //
+        // Model::preventLazyLoading(!app()->isProduction());
     }
 }