+++ database/migrations/2024_07_07_184922_create_price_categories_table.php
@@ -1,0 +1,29 @@
+<?php
+
+use Illuminate\Support\Facades\Schema;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+return new class () extends Migration {
+    /**
+     * Run the migrations.
+     */
+    public function up(): void
+    {
+        Schema::create('price_categories', function (Blueprint $table) {
+            $table->id();
+            $table->foreignId('bar_id')->constrained('bars')->onDelete('cascade');
+            $table->text('name');
+            $table->string('currency', 3);
+            $table->text('description')->nullable();
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     */
+    public function down(): void
+    {
+        Schema::dropIfExists('price_categories');
+    }
+};